+++ services/course_service_dto.go
@@ -3,5 +3,5 @@ package services
 type CreateCourseCommand struct {
 	Code  string
 	Name  string
-	Units float64
+	Units float32
 }
